Separate utils and utility
