+++ core/designsystem/src/commonMain/kotlin/com/mifos/core/designsystem/component/MifosDrawingCanvas.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.input.pointer.pointerInteropFilter
-import com.mifos.core.designsystem.utils.PathState
+import com.mifos.core.designsystem.utility.PathState
 
 @ExperimentalComposeUiApi
 @Composable
