+++ feature/collectionSheet/src/main/java/com/mifos/feature/individualCollectionSheet/individualCollectionSheet/ui/IndividualCollectionSheetScreen.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.mifos.core.designsystem.component.MifosScaffold
 import com.mifos.core.designsystem.component.MifosTabRow
 import com.mifos.core.designsystem.icon.MifosIcons
-import com.mifos.core.designsystem.utils.TabContent
+import com.mifos.core.designsystem.utility.TabContent
 import com.mifos.feature.collection_sheet.R
 import com.mifos.feature.individualCollectionSheet.newIndividualCollectionSheet.ui.NewIndividualCollectionSheetScreen
 import com.mifos.feature.individualCollectionSheet.savedIndividualCollectionSheet.ui.SavedIndividualCollectionSheetCompose
